+++ ng_trajectory/criterions/profile/profiler.py
@@ -7,6 +7,9 @@
 ######################
 
 import math, numpy
+import csv
+
+from ng_trajectory.interpolators.utils import pointsDistance
 
 from typing import List, Tuple, TextIO
 
@@ -21,6 +24,8 @@ v_0 = 0             # Initial speed [m.s^-1]
 v_lim = 4.5         # Maximum forward speed [m.s^-1]
 a_acc_max = 0.8     # Maximum longitudal acceleration [m.s^-2]
 a_break_max = 4.5   # Maximum longitudal decceleration [m.s^-2]
+_lf = 0.191         # distance from the center of mass to the front axle [m]
+_lr = 0.139         # distance from the center of mass to the rear axle [m]
 
 
 ######################
@@ -243,13 +248,75 @@ def forward_pass(points: numpy.ndarray, v_bwd: List[float], v_max: List[float], 
     return v_fwd, a, t
 
 
-def profileCompute(points: numpy.ndarray, overlap: int = 0, fd: TextIO = None, lap_time: bool = False) -> Tuple[numpy.ndarray, numpy.ndarray, numpy.ndarray]:
+def saveState(filename: str, points: numpy.ndarray, t: numpy.ndarray, v: numpy.ndarray, a: numpy.ndarray) -> bool:
+    """Save the computed profile into a CSV file.
+
+    Arguments:
+    filename -- name of the file to save state, str
+    points -- points of a trajectory with curvature, nx3 numpy.ndarray
+    v -- speed profile of the trajectory, [m.s^-1], nx1 numpy.ndarray
+    a -- acceleration profile of the trajectory, [m.s^-2], nx1 numpy.ndarray
+    t -- time of reaching the points of the trajectory, [s], nx1 or (n+1)x1 numpy.ndarray
+
+    Returns:
+    success -- True if saved, otherwise False
+
+    Note: This is inspired by the 'ng_trajectory_ros' wrapper.
+    """
+
+    x, y, k = points[:, :3].T
+    t, v, a = t.flatten(), v.flatten(), a.flatten()
+
+    # Track progress (d)
+    d = pointsDistance(points)
+    d[0] = 0.0
+    d = numpy.cumsum(d)
+
+
+    # Heading (psi)
+    psi = numpy.asarray([
+        math.atan2(
+            y[(_i + 1) % len(points)] - _y,
+            x[(_i + 1) % len(points)] - _x
+        ) for _i, (_x, _y) in enumerate(points[:, :2])
+    ])
+
+
+    with open(filename, "w", newline = "") as f:
+        writer = csv.DictWriter(f,
+            fieldnames = ["t_s", "d_m", "x_m", "y_m", "psi_rad", "k_radpm", "vx_mps", "vy_mps", "a_mps2", "omega_radps", "delta_rad"]
+        )
+
+        writer.writeheader()
+
+        for _i in range(len(points)):
+            _beta = math.atan(_lr * k[_i])
+
+            writer.writerow(
+                {
+                    "t_s": t[_i],
+                    "d_m": d[_i],
+                    "x_m": x[_i],
+                    "y_m": y[_i],
+                    "psi_rad": psi[_i],
+                    "k_radpm": k[_i],
+                    "vx_mps": v[_i] * math.cos(_beta),
+                    "vy_mps": v[_i] * math.sin(_beta),
+                    "a_mps2": a[_i],
+                    "omega_radps": v[_i] * math.cos(_beta) * k[_i],
+                    "delta_rad": (_lf + _lr) * k[_i]
+                }
+            )
+
+
+def profileCompute(points: numpy.ndarray, overlap: int = 0, fd: TextIO = None, lap_time: bool = False, save: str = None) -> Tuple[numpy.ndarray, numpy.ndarray, numpy.ndarray]:
     """Compute the speed profile using overlap.
 
     Arguments:
     points -- points of a trajectory with curvature, nx3 numpy.ndarray
     overlap -- size of trajectory overlap, int, default 0 (disabled)
     lap_time -- when True, time is appended with lap_time, bool, default False
+    save -- when given, save the solution to this file, default None (disabled)
 
     Returns:
     v_fwd -- speed profile of the trajectory after forward step, [m.s^-1],
@@ -305,6 +372,9 @@ def profileCompute(points: numpy.ndarray, overlap: int = 0, fd: TextIO = None, l
     elif lap_time:
         _t = numpy.vstack((_t, _t[0]))
         _t[0] = 0
+
+    if save is not None:
+        saveState(save, points, _t, _v, _a)
 
     return _v, _a, _t